+++ caql_test.go
@@ -1,12 +1,52 @@
 package gosnowth
 
 import (
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"strings"
 	"testing"
 )
+
+const testCAQLError = `{
+	"locals": [],
+	"method": "caql_v1",
+	"trace": [
+		"[C]:-1:error",
+		"/opt/circonus/lib/amd64/snowth/support/caql_builder.lua:133:lookup_function",
+		".../circonus/lib/amd64/snowth/support/caql_processor_v1.lua:32:node_2_pu",
+		".../circonus/lib/amd64/snowth/support/caql_processor_v1.lua:46:node_recurse",
+		".../circonus/lib/amd64/snowth/support/caql_processor_v1.lua:149:compile",
+		"/opt/circonus/lib/amd64/snowth/support/caql_fetch.lua:110:<unknown>",
+		"[C]:-1:xpcall",
+		"/opt/circonus/lib/amd64/snowth/support/utils_extension.lua:273:<unknown>",
+		"[C]:-1:xpcall",
+		"/opt/circonus/lib/amd64/snowth/support/snowth-init.lua:307:extension_process",
+		".../circonus/lib/amd64/snowth/support/luamtev-extension.lua:118:<unknown>"
+	],
+	"user_error": {
+		"message": "Function not found: histograms"
+	},
+	"status": "520 (User facing error)",
+	"arguments": {
+		"format": "DF4",
+		"q": "find:histograms(\"latency\",\"and(service:api)\")",
+		"period": 300,
+		"_id": 33545929,
+		"ignore_duration_limits": false,
+		"prepare_results": "JSON",
+		"account_id": "1",
+		"method": "fetch",
+		"start_time": 1567500000,
+		"_timeout": 15,
+		"min_prefill": 0,
+		"_debug": 0,
+		"expansion": [],
+		"end_time": 1567566000
+	},
+	"success": false
+}`
 
 func TestGetCAQLQuery(t *testing.T) {
 	ms := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter,
@@ -21,8 +61,21 @@ func TestGetCAQLQuery(t *testing.T) {
 			return
 		}
 
-		if strings.HasPrefix(r.RequestURI,
-			"/extension/lua/public/caql_v1?query=test") {
+		if r.Method == "POST" && strings.HasPrefix(r.RequestURI,
+			"/extension/lua/public/caql_v1") {
+			b, err := ioutil.ReadAll(r.Body)
+			if err != nil {
+				w.WriteHeader(http.StatusInternalServerError)
+				w.Write([]byte(testCAQLError))
+				return
+			}
+
+			if strings.Contains(string(b), "histograms") {
+				w.WriteHeader(502)
+				w.Write([]byte(testCAQLError))
+				return
+			}
+
 			w.Write([]byte(testDF4Response))
 			return
 		}
@@ -40,11 +93,12 @@ func TestGetCAQLQuery(t *testing.T) {
 	}
 
 	node := &SnowthNode{url: u}
-	res, err := sc.GetCAQLQuery(node, 1, &CAQLQuery{
-		Query:  "test",
-		Start:  0,
-		End:    900,
-		Period: 300,
+	res, err := sc.GetCAQLQuery(node, &CAQLQuery{
+		AccountID: 1,
+		Query:     "test",
+		Start:     0,
+		End:       900,
+		Period:    300,
 	})
 	if err != nil {
 		t.Fatal(err)
@@ -73,5 +127,35 @@ func TestGetCAQLQuery(t *testing.T) {
 
 	if res.Meta[0].Label != "test" {
 		t.Errorf("Expected meta label: test, got: %v", res.Meta[0].Label)
+	}
+
+	_, err = sc.GetCAQLQuery(node, &CAQLQuery{
+		AccountID: 1,
+		Query:     "find:histograms()",
+		Start:     0,
+		End:       900,
+		Period:    300,
+	})
+	if err == nil {
+		t.Fatal("Expected CAQL error response")
+	}
+
+	vErr, ok := err.(*CAQLError)
+	if !ok {
+		t.Fatalf("Expected error type: CAQLError, got: %T", err)
+	}
+
+	exp := "Function not found: histograms"
+	if vErr.UserError.Message != exp {
+		t.Errorf("Expected user error: %v, got: %v", exp,
+			vErr.UserError.Message)
+	}
+
+	exp = strings.Replace(strings.Replace(strings.Replace(testCAQLError,
+		" ", "", -1), "\t", "", -1), "\n", "", -1)
+	val := strings.Replace(strings.Replace(strings.Replace(vErr.Error(),
+		" ", "", -1), "\t", "", -1), "\n", "", -1)
+	if val != exp {
+		t.Errorf("Expected error JSON: %v, got: %v", exp, val)
 	}
 }